End work with stopnie component, add modal and topalert component

+++ assets/app/content/stopnie/stopnie.component.ts
@@ -1,7 +1,13 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, EventEmitter } from '@angular/core';
 import { StopnieService } from './stopnie.service';
 
 import { Stopien } from './add/add.model';
+
+import { ModalService } from './../../shared/modal/modal.service';
+import { TopAlertService } from './../../shared/top-alert/topalert.service';
+
+import { ModalModel, ConfirmModel } from './../../shared/modal/modal.model';
+import { TopAlertModel } from './../../shared/top-alert/topalert.model';
 
 @Component({
     selector: 'stopnie',
@@ -10,30 +16,66 @@ import { Stopien } from './add/add.model';
 export class StopnieComponent implements OnInit {
 
     list: Stopien[];
+    length: number;
+    confirm;
 
-    constructor(private StopnieService: StopnieService, ) { }
+    editEmmiter = new EventEmitter<ConfirmModel>();
 
+
+    constructor(
+        private StopnieService: StopnieService,
+        private modalService: ModalService,
+        private topAlertService: TopAlertService) { }
 
 
     ngOnInit() {
+
         this.StopnieService.list()
             .subscribe(
-                (results: Stopien[]) => {
-                     this.list = results;
-                },
-                    error => console.error(error)
+            (results: Stopien[]) => {
+                this.list = results;
+                this.length = this.list.length;
+            },
+            error => console.error(error)
             );
+
     }
 
-    delete(row){
-        this.StopnieService.delete(row.id)
+    delete(row) {
+
+        this.modalService.handleError(
+            new ModalModel(
+                "Usunięcie stopnia",
+                "Czy napewno chcesz usunąć stopień ze swojej listy",
+                "TAK, potwierdzam",
+                "NIE, zamknij"
+            ));
+
+        this.modalService.confirmResult
             .subscribe(
-                (results: Stopien[]) => {
-                     this.list = results;
-                },
-                    error => console.error(error)
+            (res: ConfirmModel) => {
+                if (res.confirm) {
+                    this.StopnieService.delete(row.id)
+                        .subscribe(
+                        () => {
+                            this.ngOnInit();
+
+                            this.topAlertService.handleAlert(
+                                new TopAlertModel(
+                                    "Stopień usunięto poprawnie",
+                                ));
+                                res.confirm = false;
+
+                        },
+                        error => console.error(error)
+                        );
+                }
+            },
+            err => console.log(err)
             );
+
     }
 
 
 }
+
